+++ test/td_dq/executions_test.exs
@@ -215,5 +215,53 @@ defmodule TdDq.ExecutionsTest do
              ] =
                Executions.list_executions(%{sources: [source_external_id, source_external_id_3]})
     end
+
+    test "list executions filtered by implementation ref" do
+      %{implementation_ref: ref} = insert(:implementation)
+      %{id: id1} = insert(:execution, implementation_ref: ref)
+      %{id: id2} = insert(:execution, implementation_ref: ref)
+
+      assert [_, _] = executions = Executions.list_executions(%{ref: ref})
+      assert_lists_equal(executions, [id1, id2], &(&1.id == &2))
+    end
+
+    test "list executions filtered by latest event status" do
+      %{id: id0} = insert(:execution)
+      %{id: id1} = insert(:execution, quality_events: [build(:quality_event, type: "FOO")])
+      %{id: id2} = insert(:execution, quality_events: [build(:quality_event, type: "BAR")])
+
+      Executions.list_executions(%{filters: %{status: ["FOO"]}})
+      |> assert_lists_equal([id1], &(&1.id == &2))
+
+      Executions.list_executions(%{filters: %{status: ["BAR"]}})
+      |> assert_lists_equal([id2], &(&1.id == &2))
+
+      Executions.list_executions(%{filters: %{status: ["FOO", "BAR"]}})
+      |> assert_lists_equal([id1, id2], &(&1.id == &2))
+
+      pending_ids =
+        %{filters: %{status: ["PENDING"]}}
+        |> Executions.list_executions()
+        |> Enum.map(& &1.id)
+
+      assert id0 in pending_ids
+      refute id1 in pending_ids
+      refute id2 in pending_ids
+    end
+  end
+
+  describe "min_max_count/1" do
+    test "returns minimum id, maximum id, and count" do
+      insert(:execution)
+      %{implementation_ref: ref} = insert(:implementation)
+
+      ids =
+        for _ <- 1..10 do
+          insert(:execution, implementation_ref: ref).id
+        end
+
+      assert %{count: 10, max_id: max_id, min_id: min_id} = Executions.min_max_count(%{ref: ref})
+      assert Enum.min_max(ids) == {min_id, max_id}
+    end
   end
 end